Visual

+++ section.php
@@ -71,8 +71,7 @@ if (! empty($action)) {
 			$pfs->load($PDOdb, GETPOST('id', 'int'));
 				$pfs->set_values($_REQUEST);
                                 
-                                
-				$pfs->save($PDOdb, GETPOST('budget'));
+                                $pfs->save($PDOdb, GETPOST('budget'), GETPOST('fk_section_parente'), GETPOST('plan_id'));
 				_card($PDOdb, $pfs, 'view');
 
 			break;
@@ -80,7 +79,11 @@ if (! empty($action)) {
 		case 'delete' :
 
 			if ($pfs->load($PDOdb, GETPOST('id', 'int'))) {
-				$pfs->delete($PDOdb);
+                            if($pfs->isDeletable($PDOdb))
+                                $pfs->delete($PDOdb);
+                            else
+                                setEventMessage($langs->trans('NotDeletable'), 'errors');
+                                
 				_list($PDOdb, $pfs);
 			} else {
 				setEventMessage($langs->trans('ImpossibleLoadElement'), 'errors');
@@ -96,11 +99,11 @@ if (! empty($action)) {
 
 			break;
                 case 'addsection' :
-                        if ($pfs->load($PDOdb, GETPOST('fk_section', 'int'))) {
+                        if ($pfs->load($PDOdb, GETPOST('fk_section_parente', 'int'))) {
 				$pfs_link = new TSectionPlanFormation();
                                 $pfs_link->fk_planform = $_REQUEST['plan_id'];
                                 $pfs_link->fk_section = GETPOST('fk_section', 'int');
-                                $pfs_link->fk_section_parente = GETPOST('id', 'int');
+                                $pfs_link->fk_section_parente = GETPOST('fk_section_parente', 'int');
                                 
                                 $pfs_link->save($PDOdb);
                                 _card($PDOdb, $pfs, 'view');
@@ -233,10 +236,12 @@ function _card(TPDOdb &$PDOdb, TSection &$pfs, $mode = '') {
 		}
 	}
         $planId = GETPOST('plan_id');
+        
         $btSave = $formCore->btsubmit($langs->trans('Valid'), 'save');
+        
         if(!empty($planId)) {
-            $btCancel = '<a class="butAction" href="' . dol_buildpath('/planformation/section.php?id=' . $pfs->rowid . '&plan_id=' . $_GET['plan_id'], 1) . '">' . $langs->trans('Cancel') . '</a>';
-            $btModifier = '<a class="butAction" href="' . dol_buildpath('/planformation/section.php?id=' . $pfs->rowid . '&plan_id=' . $_GET['plan_id'] . '&action=edit', 1) . '">' . $langs->trans('PFSectionEdit') . '</a>';
+            $btCancel = '<a class="butAction" href="' . dol_buildpath('/planformation/section.php?id=' . $pfs->rowid . '&plan_id=' .  GETPOST('plan_id'), 1) . '">' . $langs->trans('Cancel') . '</a>';
+            $btModifier = '<a class="butAction" href="' . dol_buildpath('/planformation/section.php?id=' . $pfs->rowid . '&plan_id=' . GETPOST('plan_id') . '&action=edit', 1) . '">' . $langs->trans('PFSectionEdit') . '</a>';
             $btRetour = '<a class="butAction" href="' . dol_buildpath('/planformation/planformation.php?id=' . $planId, 1) . '">' . $langs->trans('BackToList') . '</a>';
 	}
         else {
@@ -245,10 +250,7 @@ function _card(TPDOdb &$PDOdb, TSection &$pfs, $mode = '') {
             $btModifier = '<a class="butAction" href="' . dol_buildpath('/planformation/section.php?id=' . $pfs->rowid . '&action=edit', 1) . '">' . $langs->trans('PFSectionEdit') . '</a>';
         }
 	
-	
-	
-	
-	$btDelete = "<input type=\"button\" id=\"action-delete\" value=\"" . $langs->trans('Delete') . "\" name=\"cancel\" class=\"butActionDelete\" onclick=\"if(confirm('" . $langs->trans('PFDeleteConfirm') . "'))document.location.href='?action=delete&id=" . $pfs->rowid . "'\" />";
+	$btDelete = "<input type=\"button\" id=\"action-delete\" value=\"" . $langs->trans('Delete') . "\" name=\"cancel\" class=\"butActionDelete\" onclick=\"if(confirm('" . 'Etes vous sûr de supprimer ?'./*. $langs->trans('PFDeleteConfirm') .*/ "'))document.location.href='?action=delete&id=" . $pfs->rowid . "'\" />";
 
 	// Fill form with title and data
 	$data = $pfs->getTrans('title');
@@ -257,11 +259,18 @@ function _card(TPDOdb &$PDOdb, TSection &$pfs, $mode = '') {
 	
 	
         $data['budget_title'] = 'Budget';
+        $data['fk_section_parente_title'] = 'Section Parente';
         $data['plan_id'] = GETPOST('plan_id', 'int');
+		
 	if ($mode == 'edit') {
 		$data['titre'] = load_fiche_titre($pfs->getId() > 0 ? $langs->trans("PFSectionEdit") : $langs->trans("PFSectionNew"), '');
 		$data['title'] = $formCore->texte('', 'title', $pfs->title, 30, 255);
                 $data['budget'] = $formCore->texte('', 'budget', $planformSection->budget, 30, 255);
+                
+                if(GETPOST('action') !== 'new')
+                    $availableSection = $planformSection->getAvailableParentSection($PDOdb);
+                
+                $data['fk_section_parente'] = $formCore->combo("", 'fk_section_parente', $availableSection, $planformSection->fk_section_parente);
                 
 		if ($conf->global->PF_SECTION_ADDON == 'mod_planformation_section_universal') {
 			$data['ref'] = $formCore->texte('', 'ref', $pfs->ref, 15, 255);
@@ -283,6 +292,10 @@ function _card(TPDOdb &$PDOdb, TSection &$pfs, $mode = '') {
                 $data['ref'] = $formCore->texte('', 'ref', $pfs->ref, 15);
 		$data['fk_usergroup'] =  $usergroupsArray[$pfs->fk_usergroup];
 		$buttons = $btRetour . ' ' . $btModifier . ' ' . $btDelete;
+                $sectionParente = new TSection();
+                $sectionParente->load($PDOdb, $planformSection->fk_section_parente);
+                
+                $data['fk_section_parente'] = $sectionParente->title;
 	}
 	// Todo mandatory fields
 
@@ -299,20 +312,4 @@ function _card(TPDOdb &$PDOdb, TSection &$pfs, $mode = '') {
 	));
 
 	echo $formCore->end_form();
-        
-        if ($mode == 'view') {
-
-		// Combo box to add section on an other section
-		$formCore = new TFormCore($_SERVER['PHP_SELF'], 'formaddSection', 'POST');
-		echo $formCore->hidden('action', 'addsection');
-		echo $formCore->hidden('id', $pfs->getId());
-                echo $formCore->hidden('plan_id', GETPOST('plan_id', 'int'));
-		$pfsBis = new TSection();
-		$availableSection = $pfsBis->getAvailableSection($PDOdb, $pfs->getId());
-		echo $formCore->combo($langs->trans('Select Section To Add'), 'fk_section', $availableSection, '');
-		echo $formCore->btsubmit($langs->trans('Add'), 'addsection');
-		$formCore->end();
-
-		
-	}
 }
